backend/requirements.txt
fastapi
uvicorn
google-api-python-client
python-dotenv
httpx
pydantic
isodate
